Working switch2led demonstration on tic55x
This is a first release that is starting to work on real
phone hardware: Grandstream BT200.  The part that works
is the backlight to the LCD and the hook contact, but
that at least shows control.  Work has been done on the
timer actions, but that has not been made to work yet,
at least not based on interrupts.